requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=5.1.0=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
chex=0.1.3=pypi_0
click=8.1.3=pypi_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
dm-tree=0.1.7=pypi_0
docker-pycreds=0.4.0=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
faiss-gpu=1.7.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pypi_0
flatbuffers=2.0=pypi_0
flax=0.4.2=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
gdown=4.5.1=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.46.3=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.19.2=pypi_0
imgaug=0.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jax=0.3.13=pypi_0
jaxlib=0.3.10+cuda11.cudnn82=pypi_0
joblib=1.1.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_0
libgomp=11.2.0=h1234567_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pypi_0
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
msgpack=1.0.4=pypi_0
multidict=6.0.2=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.8.1=pypi_0
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h7f8727e_0
opt-einsum=3.3.0=pypi_0
optax=0.1.2=pypi_0
packaging=21.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.0.1=py310h22f2fdc_0
pip=21.2.4=py310h06a4308_0
promise=2.3=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.6.3=pypi_0
pytorch-mutex=1.0=cuda
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.8.1=pypi_0
sentry-sdk=1.6.0=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=61.2.0=py310h06a4308_0
shapely=1.8.2=pypi_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.3=hc218d9a_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.5.4=pypi_0
timm=0.6.2.dev0=pypi_0
tk=8.6.11=h1ccaba5_1
toolz=0.11.2=pypi_0
torchaudio=0.11.0=py310_cu102
torchmetrics=0.8.2=pypi_0
torchvision=0.12.0=py310_cu102
tqdm=4.64.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
wandb=0.12.19=pypi_0
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
